backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pypdf==4.3.1
python-docx==1.1.0
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
sentence-transformers==2.7.0
numpy==1.26.4
pytest==7.4.4
